:construction: refactor plugin loading
