+++ nonebot/plugin/on.pyi
@@ -1,33 +1,11 @@
 import re
-from types import ModuleType
-from dataclasses import dataclass
-from typing import Set, Dict, List, Type, Tuple, Union, Optional
+from typing import Set, List, Type, Tuple, Union, Optional
 
 from nonebot.handler import Handler
 from nonebot.matcher import Matcher
 from nonebot.permission import Permission
 from nonebot.rule import Rule, ArgumentParser
 from nonebot.typing import T_State, T_Handler, T_RuleChecker, T_StateFactory
-
-from .export import Export
-from .export import export as export
-
-plugins: Dict[str, "Plugin"] = ...
-PLUGIN_NAMESPACE: str = ...
-
-
-@dataclass(eq=False)
-class Plugin(object):
-    name: str
-    module: ModuleType
-
-    @property
-    def export(self) -> Export:
-        ...
-
-    @property
-    def matcher(self) -> Set[Type[Matcher]]:
-        ...
 
 
 def on(type: str = "",
@@ -387,40 +365,3 @@ class MatcherGroup:
             state: Optional[T_State] = ...,
             state_factory: Optional[T_StateFactory] = ...) -> Type[Matcher]:
         ...
-
-
-def load_plugin(module_path: str) -> Optional[Plugin]:
-    ...
-
-
-def load_plugins(*plugin_dir: str) -> Set[Plugin]:
-    ...
-
-
-def load_all_plugins(module_path: Set[str],
-                     plugin_dir: Set[str]) -> Set[Plugin]:
-    ...
-
-
-def load_from_json(file_path: str, encoding: str = ...) -> Set[Plugin]:
-    ...
-
-
-def load_from_toml(file_path: str, encoding: str = ...) -> Set[Plugin]:
-    ...
-
-
-def load_builtin_plugins(name: str = ...) -> Optional[Plugin]:
-    ...
-
-
-def get_plugin(name: str) -> Optional[Plugin]:
-    ...
-
-
-def get_loaded_plugins() -> Set[Plugin]:
-    ...
-
-
-def require(name: str) -> Optional[Export]:
-    ...
